Add GitHub Actions workflow for building wheels and documentation for installation and usage Don't allow keep_alive or call_guard on properties (#5533)
* Don't allow keep_alive or call_guard on properties

The def_property family blindly ignore the keep_alive and call_guard arguments passed to them making them confusing to use.
This adds a static_assert if either is passed to make it clear it doesn't work.
I would prefer this to be a compiler warning but I can't find a way to do that. Is that even possible?

* style: pre-commit fixes

* Re-run tests

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>